Integration user+produit+blogs

+++ src/Controller/CommentaireController.php
@@ -15,12 +15,19 @@ class CommentaireController extends AbstractController
     #[Route('/commentaire/{id}/edit', name: 'app_commentaire_edit', methods: ['GET'])]
     public function getCommentaire(Commentaire $commentaire): JsonResponse
     {
+        if ($this->getUser() !== $commentaire->getUtilisateur() && !$this->isGranted('ROLE_ADMIN')) {
+            return $this->json(['success' => false, 'message' => 'Non autorisé'], 403);
+        }
         return $this->json(['contenu' => $commentaire->getContenu()]);
     }
 
     #[Route('/commentaire/{id}/update', name: 'app_commentaire_update', methods: ['POST'])]
     public function update(Request $request, Commentaire $commentaire, EntityManagerInterface $em): JsonResponse
     {
+        if ($this->getUser() !== $commentaire->getUtilisateur() && !$this->isGranted('ROLE_ADMIN')) {
+            return $this->json(['success' => false, 'message' => 'Non autorisé'], 403);
+        }
+
         $data = json_decode($request->getContent(), true);
         if (isset($data['contenu'])) {
             $commentaire->setContenu($data['contenu']);
@@ -34,6 +41,11 @@ class CommentaireController extends AbstractController
     public function delete(Request $request, Commentaire $commentaire, EntityManagerInterface $em): Response
     {
         $articleId = $commentaire->getArticle()->getId();
+
+        if ($this->getUser() !== $commentaire->getUtilisateur() && !$this->isGranted('ROLE_ADMIN')) {
+             return $this->redirectToRoute('app_article_show', ['id' => $articleId]);
+        }
+
         if ($this->isCsrfTokenValid('delete' . $commentaire->getId(), $request->request->get('_token'))) {
             $em->remove($commentaire);
             $em->flush();
